# Speedup from_lc further