--- a/docs/changes/757.trivial.rst
+++ b/docs/changes/757.trivial.rst
@@ -1,0 +1,1 @@
+Speedup creation of events in ``EventList.from_lc``