--- a/setup.cfg
+++ b/setup.cfg
@@ -101,6 +101,7 @@ filterwarnings =
     ignore:Large Datasets may not be processed efficiently:UserWarning
     ignore:.*is a deprecated alias for:DeprecationWarning
     ignore:.*HIERARCH card will be created.*:
+    ignore:.*FigureCanvasAgg is non-interactive.*:UserWarning
 
 ;addopts = --disable-warnings